Huge update
 - added a quick pmem emulator script (simply mounts dram as pmem)
    - check 'emulating pmem.md' for more information

 - cleaned up APIs in both volatile and persistent stms

 - fixed broken nesting logic
    - (open) nesting doesn't need to wait for lock snapshot to become
    even when begining in an inner transaction
    - only the outer one needs to do this
    - also since we decrement tx nesting level in the end function, we
    don't need the preabort functions

 - renamed persistent version of stm apis to PTM_*
 - added 3 different tests for tm_hashmap, bench, test and transfer
    - consequently had to unify their Makefiles
    - subsequently had to battle with make once more :(

 - fixed insanely broken hashmap resizing logic
    - spent perhaps too long on this...

 - added non-tx put method for hashmap along with non-tx resize
 - added zalloc functions for all 4 algs
 - added local reading STM/PTM wrappers along with (struct) field READ
 and WRITE helper functions

+++ pmdk_stms/algs/norec/norec_util.c
@@ -86,7 +86,7 @@ void tx_vector_empty(struct tx_vec *vec)
 #define MAX_GROWS 8
 #define INIT_CAP TABLE_PRIMES[0]
 
-static int TABLE_PRIMES[] = {11, 23, 53, 97, 193, 389, 769, 1543, 3079};
+static size_t TABLE_PRIMES[] = {11, 23, 53, 97, 193, 389, 769, 1543, 3079};
 
 uint64_t _hash(uintptr_t key)
 {
